Implement first part of authorization code flow
This creates two new endpoints (or changes the old, test endpoints) with
the help of oauthlib. It currently implements the standard authorization
endpoint described in [section 3.1 in RFC 6749][rfc-authz-endpoint].

`GET /oauth/authorize`
----------------------

This endpoint validates the data in the request parameters, typically a
`client_id`, `redirect_uri`, and `response_type`. These parameters are
checked againts the permissions stored on the `AuthClient` model.
When the request is valid it shows an authorize screen to the logged-in
user with information about the OAuth client requesting access to the
users' data.

`POST /oauth/authorize`
-----------------------

This endpoint is called when the user accepts the authorize screen that
is rendered in the previous endpoint.
The endpoint checks again that the requested data is consistent with
the permissions stored in the `AuthClient` model, and creates an
authorization code with a default set of scopes (`annotation:read`,
`annotation:write`), and then redirects the browser to the client's
preferred `redirect_uri` with the authorization code.

[rfc-authz-endpoint]: https://tools.ietf.org/html/rfc6749#section-3.1